rename browser_utils to screen_parsing
