+++ setup.py
@@ -45,7 +45,11 @@ setuptools.setup(
     long_description_content_type="text/markdown",
     url="https://github.com/microsoft/autogen",
     packages=setuptools.find_packages(include=["autogen*"], exclude=["test"]),
-    package_data={"autogen.agentchat.contrib.multimodal_web_surfer": ["*.*"]},
+    package_data={
+        "autogen.agentchat.contrib.multimodal_web_surfer": ["*.*"],
+        'autogen.screen_parsing': ['autogen/screen_parsing/static/*']
+    },
+    include_package_data=True,
     install_requires=install_requires,
     extras_require={
         "test": [
